Splitting function tests for item into smaller files. i.e. batch, bulk, item
