+++ sdk/cosmosdb/cosmos/tsconfig.strict.json
@@ -160,8 +160,10 @@
     "test/public/functional/database.spec.ts",
     "test/public/functional/databaseaccount.spec.ts",
     "test/public/functional/globalEndpointManager.spec.ts",
-    "test/public/functional/item.spec.ts",
-    "test/public/functional/itemIdEncoding.spec.ts",
+    "test/public/functional/item/item.spec.ts",
+    "test/public/functional/item/bulk.item.spec.ts",
+    "test/public/functional/item/batch.item.spec.ts",
+    "test/public/functional/item/itemIdEncoding.spec.ts",
     "test/public/functional/npcontainer.spec.ts",
     "test/public/functional/offer.spec.ts",
     "test/public/functional/permission.spec.ts",
